pseudo_label

+++ fish/effdet/pseudo_label.py
@@ -1,6 +1,6 @@
 import glob, tqdm, torch, json, shutil
 from pathlib import Path
-from typing import Dict, Any
+from typing import Dict, Any, List
 from torch.utils.data import DataLoader
 from object_detection.utils import DetectionPlot
 from object_detection.entities.box import (
@@ -13,28 +13,38 @@ from object_detection.entities.box import (
     resize,
     filter_size,
     resize,
+    box_hflip,
+    box_vflip,
 )
 from fish.data import (
     read_test_rows,
-    TestDataset,
+    LabeledDataset,
     test_transforms,
     inv_normalize,
-    add_submission,
-    Submission,
+    annotate,
 )
-from fish.centernet import config
-from fish.centernet.train import model, model_loader, to_boxes
+from fish.effdet import config
+from fish.effdet.train import model, model_loader, to_boxes, collate_fn
 from ensemble_boxes import weighted_boxes_fusion
 from torchvision.transforms.functional import hflip, vflip
+from fish.store import StoreApi, Box
+from toolz.curried import pipe, filter
+
+Submission = Dict[str, Any]
+
+
+store = StoreApi()
 
 
 @torch.no_grad()
 def predict(device: str) -> None:
-    rows = read_test_rows("/store")
-    out_dir = Path("/store/submission")
-    shutil.rmtree(out_dir)
+    rows = store.filter()
+    rows = pipe(rows, filter(lambda x: "test" in x['id']), list)
+    out_dir = Path("/store/pseudo")
+    if out_dir.exists():
+        shutil.rmtree(out_dir)
     out_dir.mkdir(exist_ok=True)
-    dataset = TestDataset(rows=rows, transforms=test_transforms)
+    dataset = LabeledDataset(rows=rows, transforms=test_transforms)
     net = model_loader.load_if_needed(model).to(device).eval()
     loader = DataLoader(
         dataset,
@@ -42,17 +52,16 @@ def predict(device: str) -> None:
         num_workers=1,
         shuffle=False,
         drop_last=False,
+        collate_fn=collate_fn
     )
     weights = [1, 1]
-    submission: Submission = {}
-    for ids, image_batch in tqdm.tqdm(loader):
+    for image_batch, _, _, ids in tqdm.tqdm(loader):
         image_batch = image_batch.to(device)
-
+        _, _, h, w = image_batch.shape
+        box_batch, confidence_batch, label_batch = to_boxes(net(image_batch))
         h_box_batch, h_confidence_batch, h_label_batch = to_boxes(
             net(hflip(image_batch))
         )
-        box_batch, confidence_batch, label_batch = to_boxes(net(image_batch))
-
         for (
             img,
             id,
@@ -72,33 +81,29 @@ def predict(device: str) -> None:
             label_batch,
             h_label_batch,
         ):
-            _, h, w = img.shape
-
-            _boxes, confidences, labels = weighted_boxes_fusion(
+            m_boxes, m_confidences, m_labels = weighted_boxes_fusion(
                 [
-                    yolo_to_pascal(boxes, (1, 1)),
-                    yolo_to_pascal(yolo_hflip(h_boxes), (1, 1)),
+                    resize(boxes, (1 / w, 1 / h)),
+                    box_hflip(resize(h_boxes, (1 / w, 1 / h)), (1, 1)),
                 ],
                 [confidences, h_confidences],
                 [labels, h_labels],
                 iou_thr=config.iou_threshold,
                 weights=weights,
-                skip_box_thr=config.to_boxes_threshold,
             )
-            _boxes = resize(PascalBoxes(torch.from_numpy(_boxes)), (w, h))
-            filter_indices = confidences > config.to_boxes_threshold
-            _boxes = _boxes[filter_indices]
-            labels = labels[filter_indices]
-            plot = DetectionPlot(inv_normalize(img))
-            plot.draw_boxes(boxes=_boxes, confidences=confidences, labels=labels)
-            plot.save(out_dir.joinpath(f"{id}.jpg"))
-            _boxes = resize(
-                _boxes, scale=(config.original_width / w, config.original_height / h)
-            )
-            add_submission(submission, id, boxes=_boxes, labels=labels)
 
-    with open(out_dir.joinpath("submission.json"), "w") as f:
-        json.dump(submission, f)
+            m_confidences = torch.from_numpy(m_confidences)
+            indices = m_confidences > config.pseudo_threshold
+            m_boxes = PascalBoxes(torch.from_numpy(m_boxes)[indices])
+            m_labels = Labels(torch.from_numpy(m_labels)[indices])
+            m_confidences = Confidences(m_confidences[indices])
+            print(m_confidences)
+
+            plot = DetectionPlot(inv_normalize(img))
+            plot.draw_boxes(boxes=resize(m_boxes, (w,h)), labels=m_labels, confidences=m_confidences)
+            plot.save(out_dir.joinpath(f"{id}.jpg"))
+
+            annotate(store, id, boxes=m_boxes, labels=m_labels)
 
 
 if __name__ == "__main__":
